va queriendo

+++ C_entidad/DetalleCompra.cs
@@ -9,9 +9,9 @@ namespace FankyRecords.C_entidad
     public class DetalleCompra
     {
         public int ID_detalleCompra { get; set; }
-        public int PrecioCompra { get; set; }
+        public decimal PrecioCompra { get; set; }
         public int Cantidad { get; set; }
-        public float SubTotal { get; set; }
+        public decimal SubTotal { get; set; }
         public Productos Obj_producto { get; set; }
         public RegistrarCompra Obj_registrarCompra { get; set; }
